requirements.txt
aiofiles==0.7.0
alembic==1.7.3
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
fastapi==0.68.1
FastAPI-SQLAlchemy==0.2.1
flower==1.0.0
greenlet==1.1.1
h11==0.12.0
humanize==3.11.0
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.5
MarkupSafe==2.0.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
tornado==6.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
